+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## [Unreleased]
 
 ### Fixed
-- Fixed issue with filter editor column selector generating react errors due to missing key in array rendering. (#48)
+- Fixed issue with filter editor column selector generating react errors due to missing key in array rendering (#48)
+- Fixed issue with filter default values for multi-valued operators (#43)
 
 ## [1.5.1] - 2022-01-24
 ### Added